+++ ui/litellm-dashboard/eslint-suppressions.json
@@ -2272,17 +2272,17 @@
       "count": 1
     }
   },
-  "src/components/tag_management/TagTable.tsx": {
+  "src/app/(dashboard)/tag-management/_components/TagTable.tsx": {
     "no-restricted-imports": {
       "count": 1
     }
   },
-  "src/components/tag_management/components/CreateTagModal.tsx": {
+  "src/app/(dashboard)/tag-management/_components/components/CreateTagModal.tsx": {
     "no-restricted-imports": {
       "count": 1
     }
   },
-  "src/components/tag_management/index.tsx": {
+  "src/app/(dashboard)/tag-management/_components/index.tsx": {
     "no-restricted-imports": {
       "count": 1
     },
@@ -2290,7 +2290,7 @@
       "count": 1
     }
   },
-  "src/components/tag_management/tag_info.tsx": {
+  "src/app/(dashboard)/tag-management/_components/tag_info.tsx": {
     "no-restricted-imports": {
       "count": 1
     },
@@ -2388,12 +2388,12 @@
       "count": 2
     }
   },
-  "src/components/vector_store_management/CreateVectorStore.tsx": {
+  "src/app/(dashboard)/vector-stores/_components/CreateVectorStore.tsx": {
     "no-restricted-imports": {
       "count": 1
     }
   },
-  "src/components/vector_store_management/VectorStoreForm.tsx": {
+  "src/app/(dashboard)/vector-stores/_components/VectorStoreForm.tsx": {
     "no-nested-ternary": {
       "count": 2
     },
@@ -2404,12 +2404,12 @@
       "count": 1
     }
   },
-  "src/components/vector_store_management/VectorStoreTable.tsx": {
+  "src/app/(dashboard)/vector-stores/_components/VectorStoreTable.tsx": {
     "no-restricted-imports": {
       "count": 1
     }
   },
-  "src/components/vector_store_management/index.tsx": {
+  "src/app/(dashboard)/vector-stores/_components/index.tsx": {
     "no-restricted-imports": {
       "count": 1
     },
@@ -2417,7 +2417,7 @@
       "count": 1
     }
   },
-  "src/components/vector_store_management/vector_store_info.tsx": {
+  "src/app/(dashboard)/vector-stores/_components/vector_store_info.tsx": {
     "no-restricted-imports": {
       "count": 1
     },
