+++ ui/litellm-dashboard/src/app/(dashboard)/tag-management/_components/tag_info.tsx
@@ -12,13 +12,13 @@ import {
 } from "@tremor/react";
 import { Form, Input, Select as Select2, Tooltip } from "antd";
 import { InfoCircleOutlined } from "@ant-design/icons";
-import { fetchUserModels } from "../organisms/create_key_button";
-import { getModelDisplayName } from "../key_team_helpers/fetch_available_models_team_key";
-import { tagInfoCall, tagUpdateCall } from "../networking";
-import { Tag } from "./types";
-import NotificationsManager from "../molecules/notifications_manager";
-import NumericalInput from "../shared/numerical_input";
-import BudgetDurationDropdown from "../common_components/budget_duration_dropdown";
+import { fetchUserModels } from "@/components/organisms/create_key_button";
+import { getModelDisplayName } from "@/components/key_team_helpers/fetch_available_models_team_key";
+import { tagInfoCall, tagUpdateCall } from "@/components/networking";
+import { Tag } from "@/components/tag_management/types";
+import NotificationsManager from "@/components/molecules/notifications_manager";
+import NumericalInput from "@/components/shared/numerical_input";
+import BudgetDurationDropdown from "@/components/common_components/budget_duration_dropdown";
 import { copyToClipboard as utilCopyToClipboard } from "@/utils/dataUtils";
 import { CheckIcon, CopyIcon } from "lucide-react";
 import { Button as AntdButton } from "antd";
