+++ ui/litellm-dashboard/src/app/(dashboard)/vector-stores/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import VectorStoreManagement from "@/components/vector_store_management";
+import VectorStoreManagement from "./_components";
 import useAuthorized from "@/app/(dashboard)/hooks/useAuthorized";
 
 export default function VectorStores() {
