+++ ui/litellm-dashboard/tests/CreateKeyPage.expiredToken.test.tsx
@@ -131,8 +131,8 @@ vi.mock("@/components/guardrails", () => ({ default: stub("guardrails") }));
 vi.mock("@/components/prompts", () => ({ default: stub("prompts") }));
 vi.mock("@/components/transform_request", () => ({ default: stub("transform-request") }));
 vi.mock("@/components/mcp_tools", () => ({ MCPServers: stub("mcp-servers") }));
-vi.mock("@/components/tag_management", () => ({ default: stub("tag-management") }));
-vi.mock("@/components/vector_store_management", () => ({ default: stub("vector-stores") }));
+vi.mock("@/app/(dashboard)/tag-management/_components", () => ({ default: stub("tag-management") }));
+vi.mock("@/app/(dashboard)/vector-stores/_components", () => ({ default: stub("vector-stores") }));
 vi.mock("@/components/ui_theme_settings", () => ({ default: stub("ui-theme-settings") }));
 vi.mock("@/components/organisms/create_key_button", () => ({ fetchUserModels: vi.fn() }));
 vi.mock("@/components/common_components/fetch_teams", () => ({ fetchTeams: vi.fn() }));
